backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.0.2
sqlalchemy==2.0.23
pydantic==2.5.0
discord.py==2.3.2